fix build fail

+++ wmserver/src/window_manager_proxy.cpp
@@ -361,7 +361,8 @@ WMError WindowManagerProxy::SetWindowAnimationController(const sptr<RSIWindowAni
         return WMError::WM_ERROR_IPC_FAILED;
     }
 
-    auto error = Remote()->SendRequest(WindowManagerMessage::TRANS_ID_ANIMATION_SET_CONTROLLER, data, reply, option);
+    auto error = Remote()->SendRequest(static_cast<uint32_t>(WindowManagerMessage::TRANS_ID_ANIMATION_SET_CONTROLLER),
+        data, reply, option);
     if (error != ERR_NONE) {
         WLOGFE("Send request error: %{public}d", error);
         return WMError::WM_ERROR_IPC_FAILED;
